TypeScript Support
