+++ example/lib/onCreateNode.ts
@@ -1,0 +1,15 @@
+import { createFilePath } from "gatsby-source-filesystem"
+import { GatsbyNode } from "gatsby"
+
+export const onCreateNode: GatsbyNode["onCreateNode"] = ({ node, actions, getNode }) => {
+  const { createNodeField } = actions
+
+  if (node.internal.type === `MarkdownRemark`) {
+    const value = createFilePath({ node, getNode })
+    createNodeField({
+      name: `slug`,
+      node,
+      value,
+    })
+  }
+}
